requirements.txt
fastapi
uvicorn
requests
websockets
Pillow
pydantic
python-dotenv
aiohttp
